Fix a bug where run lengths over 251 would crash
If the compressed data encoded a run length over 251, the program would
crash upon decompressing when it came to a step where it tried to add 4
to the run length to count how many bytes it should emit into the
decompressed output. This fixes the bug by guarding against the size of
the run length and emitting an error if it is too large.