implement `peek_n_bits()`
We implemented the `peek_n_bits()` function and did a bunch of
refactoring to use it in various places. We also added tests and
refactored the `bitstream` module to use `io::Error` exclusively.

We wrote `bits_in_buffer()` and this took us quite a long time.

Next time: Look for the TODONEXT comment! We should write
`peek_integer()` and then finally use it to determine if we're at the
end of the stream or if there's another block to be parsed.