We finished parsing the file
We can now get symbols out of the file which we can run through the
decompression pipeline. We have a failing unit test now due to an issue
with the origin pointer. We noticed that we weren't using the origin
pointer field in the header, which seems like a likely culprit. We
should investigate that next time.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>