use the origin pointer
We rectified our questionable decision to append the origin pointer to
the end of the input data in the decode pipeline. We pass the origin
pointer through the pipeline as an integer. We actually used the origin
pointer we parsed out of the file, and we make it all the way through
our decode pipeline! We don't get the data we wanted, but the
shareholders decided that wasn't critical.

TODONEXT: Overthrow the shareholders and show them the error of their
ways. Debug why the `can_we_read` test is decompressing all ones.